Implemented the authentication. Also implemented one more custom exception.

+++ src/main/java/com/company/web/springdemo/services/BeerServiceImpl.java
@@ -2,7 +2,9 @@ package com.company.web.springdemo.services;
 
 import com.company.web.springdemo.exceptions.EntityDuplicateException;
 import com.company.web.springdemo.exceptions.EntityNotFoundException;
+import com.company.web.springdemo.exceptions.UnauthorizedOperationException;
 import com.company.web.springdemo.models.Beer;
+import com.company.web.springdemo.models.User;
 import com.company.web.springdemo.repositories.BeerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -46,7 +48,10 @@ public class BeerServiceImpl implements BeerService {
     }
 
     @Override
-    public void update(Beer beer) {
+    public void update(Beer beer, User user) {
+        if (!user.isAdmin()) {
+            throw new UnauthorizedOperationException("Only admins or owners can modify the beer.");
+        }
         boolean duplicateExists = true;
         try {
             Beer existingBeer = repository.get(beer.getName());
@@ -65,7 +70,10 @@ public class BeerServiceImpl implements BeerService {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(int id, User user) {
+        if (!user.isAdmin()) {
+            throw new UnauthorizedOperationException("Only admins or owners can delete the beer.");
+        }
         repository.delete(id);
     }
